requirements.txt
Django
factory_boy
six
